Fix authz/resource-server/policy/... endpoints
Closes: #26851
